[11.0][MIG] hr_expense_sequence: backport from v12 to v11
