+++ hr_expense_sequence/__manifest__.py
@@ -4,7 +4,7 @@
 
 {
     'name': 'HR expense sequence',
-    'version': '12.0.1.0.0',
+    'version': '11.0.1.0.0',
     'license': 'AGPL-3',
     'category': 'Human Resources',
     'author': 'Serv. Tecnol. Avanzados - Pedro M. Baeza,'
